requirements.txt
fastapi[standard]
pydantic
torch
torchvision
Pillow